backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
sendgrid
apscheduler
python-dotenv